Manage external tools via .chezmoiexternals
Relying on scripts for downloading binaries is not longer necessary
after the introduction of "archive-file" in chezmoi. This breaks the
building process for custom arch packages that are in this repository.
It's by intent, cause that packaging process was broken anyway and needs
to be replaced with a better solution asap, which is out of the scope
for this commit.

Breaking change: custom Arch packages are no longer built on 'cz update'.