jj: several optimizations
While most of the things are just cosmetic, I do not longer hardcode
the value for the trunk() revset. This is for one specific reason: Newer
repositories cloned with jj automatically have it set and for older
ones, the value was actually not behaving as intended.

In order to reduce my maintenance work, I decided to get rid of the
default.