Add little script to rewrite "message:" links for Betterbird
RFC2392 defines "mid" as the recommended scheme for links to message
IDs. Unfortunately, Apple derives from this standard and introduced
their own "message" scheme. Fortunately for us, it's just the ID of the
message with no additional encoding or obscurity.

To achieve a seamless workflow between by iPhone and the desktop, a
little fish script got added to rewrite the URI. Because the handler
needs to be either in PATH or defined with the full path to the
executable, I leverage the templating functionality of chezmoi here to
ensure portability.